requirements.txt
customtkinter>=5.2.1
Pillow>=10.2.0
reportlab>=4.0.7
openpyxl>=3.1.2
